fixed transparent block lighting

+++ src/overviewer.h
@@ -41,7 +41,7 @@ Imaging imaging_python_to_c(PyObject *obj);
 PyObject *alpha_over(PyObject *dest, PyObject *src, PyObject *mask, int dx,
                      int dy, int xsize, int ysize);
 PyObject *alpha_over_wrap(PyObject *self, PyObject *args);
-PyObject *brightness(PyObject *img, double factor);
+PyObject *brightness(PyObject *img, float factor);
 
 /* in iterate.c */
 typedef struct {
